Add authorization checks for TourLog create, delete, update

+++ tourplanner-backend/src/main/java/com/example/tourplannerbackend/service/TourLogService.java
@@ -26,23 +26,37 @@ public class TourLogService {
         return tourLogRepository.findAll();
     }
 
-    public TourLog createTourLog(Long tourId, TourLog tourLog){
+    public TourLog createTourLog(Long tourId, TourLog tourLog, String username){
         Tour tour = tourRepository.findById(tourId).orElseThrow(()-> new RuntimeException("Tour nciht gefunden"));
+
+        if (!tour.getUser().getUsername().equals(username)) {
+            throw new RuntimeException("Unauthorized");
+        }
+
         tourLog.setTour(tour);
-        logger.info("TourLog created for tour: {}", tourId);
+        logger.info("TourLog created for tour: {} by: {}", tourId, username);
         return tourLogRepository.save(tourLog);
     }
 
-    public void deleteTourLog(Long id){
+    public void deleteTourLog(Long id, Long tourId, String username){
+
+        Tour tour = tourRepository.findById(tourId).orElseThrow(()-> new RuntimeException("Tour nciht gefunden"));
+        if (!tour.getUser().getUsername().equals(username)) {
+            throw new RuntimeException("Unauthorized");
+        }
         tourLogRepository.deleteById(id);
-        logger.info("TourLog deleted: {}", id);
+        logger.info("TourLog deleted: {} by: {}", id, username);
     }
 
-    public TourLog updateTourLog(Long id, TourLog updatedLog){
+    public TourLog updateTourLog(Long id, TourLog updatedLog, String username){
         TourLog existing = tourLogRepository.findById(id).orElseThrow();
+        if (!existing.getTour().getUser().getUsername().equals(username)) {
+            throw new RuntimeException("Unauthorized");
+        }
+
         updatedLog.setId(id);
         updatedLog.setTour(existing.getTour());
-        logger.info("TourLog updated: {}", id);
+        logger.info("TourLog updated: {} by: {}", id, username);
         return tourLogRepository.save(updatedLog);
     }
